requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-huggingface
faiss-cpu
sentence-transformers
pydantic
python-dotenv
typing-extensions
langchain-openai
langchain-ollama
